updated hello world. (removed the duplicated code).
I unfortunately was unable to figure out how to create a while loop for this assignment. I will be reviewing chapter 2 on Saturday to see if i can make it happen then. I saw your previous comment about the code being duplicated. This was done so that when opening the program in the Python shell it would loop the initial choose your language prompt . I have since changed the code back to comply with the assigned tasks let me know if everything looks ok please and thank you.